d/control: Avoid libwacom-dev when not using xserver-xorg-input-wacom
The upstream build system uses the presence or absence of libwacom-dev
to decide whether to build the Wacom plugin, which requires
xorg-wacom.pc from xserver-xorg-input-wacom. That package is unavailable
on s390x (and presumably Hurd and kFreeBSD), causing build failure.

Avoid this by dropping the build-dependency on libwacom-dev on
architectures where xserver-xorg-input-wacom can't be installed,
and moving it to Build-Conflicts on those architectures.

Closes: #977762